+++ examples/dynamic_fanout.py
@@ -24,11 +24,11 @@ def consume(value: int):
 
 # TODO: replace the domain and token with your own
 ws = WorkflowService(host='https://my-argo-server.com', token='my-auth-token')
-w = Workflow('dynamic-fan-out', ws)
+w = Workflow('dynamic-fanout', ws)
 generate_task = Task('generate', generate)
 consume_task = Task('consume', consume, input_from=InputFrom(name='generate', parameters=['value']))
 
 generate_task.next(consume_task)
 
 w.add_tasks(generate_task, consume_task)
-w.submit()
+w.create()