+++ examples/k8s_sa.py
@@ -47,7 +47,7 @@ token = get_sa_token("argo-server", namespace=namespace)
 
 # TODO: replace the domain and token with your own
 ws = WorkflowService(host='https://my-argo-server.com', token='my-auth-token', namespace=namespace)
-w = Workflow("hello-hera", ws)
+w = Workflow("k8s-sa", ws)
 t = Task("t", hello)
 w.add_task(t)
 ws.submit(w.workflow, namespace=namespace)