+++ src/hera/cron_workflow.py
@@ -2,7 +2,6 @@
 from datetime import datetime
 from datetime import timezone as tz
 from typing import Dict, List, Optional, Tuple
-from uuid import uuid4
 
 import pytz
 from argo_workflows.models import (
@@ -21,7 +20,7 @@ from hera.cron_workflow_service import CronWorkflowService
 from hera.security_context import WorkflowSecurityContext
 from hera.task import Task
 from hera.ttl_strategy import TTLStrategy
-from hera.volumes import Volume
+from hera.workflow_editors import add_head, add_tail, add_task, add_tasks
 
 
 class CronWorkflow:
@@ -49,7 +48,7 @@ class CronWorkflow:
     annotations: Optional[Dict[str, str]] = None
         A Dict of annotations to attach to the CronWorkflow object metadata.
     namespace: Optional[str] = 'default'
-        The namespace to use by default when calling create/suspend/resume.  Defaults to 'default'.
+        The namespace to use by default when calling create/update/suspend/resume.  Defaults to 'default'.
     security_context:  Optional[WorkflowSecurityContext] = None
         Define security settings for all containers in the workflow.
     image_pull_secrets: Optional[List[str]] = None
@@ -81,7 +80,7 @@ class CronWorkflow:
         if timezone and timezone not in pytz.all_timezones:
             raise ValueError(f'{timezone} is not a valid timezone')
 
-        self.name = f'{name.replace("_", "-")}-{str(uuid4()).split("-")[0]}'
+        self.name = name.replace("_", "-")
         self.schedule = schedule
         self.timezone = timezone
         self.service = service
@@ -149,82 +148,28 @@ class CronWorkflow:
         )
 
     def add_task(self, t: Task) -> None:
-        """Adds a single task to the workflow"""
-        self.add_tasks(t)
+        add_task(self, t)
 
     def add_tasks(self, *ts: Task) -> None:
-        """Adds multiple tasks to the workflow"""
-        if not all(ts):
-            return
-
-        for t in ts:
-            self.spec.templates.append(t.argo_template)
-
-            if t.resources.volumes:
-                for vol in t.resources.volumes:
-                    if isinstance(vol, Volume):
-                        # dynamically provisioned volumes need associated claims on the workflow spec
-                        self.spec.volume_claim_templates.append(vol.get_claim_spec())
-                    else:
-                        self.spec.volumes.append(vol.get_volume())
-
-            self.dag_template.tasks.append(t.argo_task)
+        add_tasks(self, *ts)
 
     def add_head(self, t: Task, append: bool = True) -> None:
-        """Adds a task at the head of the workflow so the workflow start with the given task.
-
-        This sets the given task as a dependency of the starting tasks of the workflow.
-
-        Parameters
-        ----------
-        t: Task
-            The task to add to the head of the workflow.
-        append: bool = True
-            Whether to append the given task to the workflow.
-        """
-        if append:
-            self.add_task(t)
-
-        for template_task in self.dag_template.tasks:
-            if template_task.name != t.name:
-                if hasattr(template_task, 'dependencies'):
-                    template_task.dependencies.append(t.name)
-                else:
-                    setattr(template_task, 'dependencies', [t.name])
+        add_head(self, t, append=append)
 
     def add_tail(self, t: Task, append: bool = True) -> None:
-        """Adds a task as the tail of the workflow so the workflow ends with the given task.
-
-        This sets the given task's dependencies to all the tasks that are not listed as dependencies in the workflow.
-
-        Parameters
-        ----------
-        t: Task
-            The task to add to the tail of the workflow.
-        append: bool = True
-            Whether to append the given task to the workflow.
-        """
-        if append:
-            self.add_task(t)
-
-        dependencies = set()
-        task_name_to_task = dict()
-        for template_task in self.dag_template.tasks:
-            if hasattr(template_task, 'dependencies'):
-                dependencies.update(template_task.dependencies)
-            if template_task.name != t.name:
-                task_name_to_task[template_task.name] = template_task
-
-        # the tasks that are not listed as dependencies are "end tasks" i.e nothing follows after
-        # e.g if A -> B -> C then B.deps = [A] and C.deps = [B] but nothing lists C so C is "free"
-        free_tasks = set(task_name_to_task.keys()).difference(dependencies)
-        t.argo_task.dependencies = list(free_tasks)
+        add_tail(self, t, append=append)
 
     def create(self, namespace: Optional[str] = None) -> IoArgoprojWorkflowV1alpha1CronWorkflow:
         """Creates the cron workflow in the server"""
         if namespace is None:
             namespace = self.namespace
         return self.service.create(self.workflow, namespace)
+
+    def update(self, namespace: Optional[str] = None) -> IoArgoprojWorkflowV1alpha1CronWorkflow:
+        """Updates the cron workflow in the server"""
+        if namespace is None:
+            namespace = self.namespace
+        return self.service.update(self.workflow, namespace)
 
     def suspend(self, name: Optional[str] = None, namespace: Optional[str] = None) -> Tuple[object, int, dict]:
         """Suspends the cron workflow"""