+++ src/hera/ttl_strategy.py
@@ -1,7 +1,7 @@
-from pydantic import BaseModel
 from typing import Optional
 
 from argo_workflows.models import IoArgoprojWorkflowV1alpha1TTLStrategy
+from pydantic import BaseModel
 
 
 class TTLStrategy(BaseModel):