+++ src/hera/workflow.py
@@ -1,7 +1,5 @@
 """The implementation of a Hera workflow for Argo-based workflows"""
-import warnings
-from typing import Dict, List, Optional
-from uuid import uuid4
+from typing import Dict, List, Optional, Tuple
 
 from argo_workflows.models import (
     IoArgoprojWorkflowV1alpha1DAGTemplate,
@@ -16,12 +14,8 @@ from argo_workflows.models import (
 from hera.security_context import WorkflowSecurityContext
 from hera.task import Task
 from hera.ttl_strategy import TTLStrategy
-from hera.volumes import Volume
+from hera.workflow_editors import add_head, add_tail, add_task, add_tasks
 from hera.workflow_service import WorkflowService
-
-# by default, `DeprecationWarning`s are silenced, this removes the warning from the filter so it
-# can be issued to users
-warnings.simplefilter('always', DeprecationWarning)
 
 
 class Workflow:
@@ -74,7 +68,7 @@ class Workflow:
         workflow_template_ref: Optional[str] = None,
         ttl_strategy: Optional[TTLStrategy] = None,
     ):
-        self.name = f'{name.replace("_", "-")}-{str(uuid4()).split("-")[0]}'  # RFC1123
+        self.name = f'{name.replace("_", "-")}'  # RFC1123
         self.namespace = namespace or 'default'
         self.service = service
         self.parallelism = parallelism
@@ -133,91 +127,25 @@ class Workflow:
         self.workflow = IoArgoprojWorkflowV1alpha1Workflow(metadata=self.metadata, spec=self.spec)
 
     def add_task(self, t: Task) -> None:
-        """Adds a single task to the workflow"""
-        self.add_tasks(t)
+        add_task(self, t)
 
     def add_tasks(self, *ts: Task) -> None:
-        """Adds multiple tasks to the workflow"""
-        if not all(ts):
-            return
-
-        for t in ts:
-            self.spec.templates.append(t.argo_template)
-
-            if t.resources.volumes:
-                for vol in t.resources.volumes:
-                    if isinstance(vol, Volume):
-                        # dynamically provisioned volumes need associated claims on the workflow spec
-                        self.spec.volume_claim_templates.append(vol.get_claim_spec())
-                    else:
-                        self.spec.volumes.append(vol.get_volume())
-
-            self.dag_template.tasks.append(t.argo_task)
+        add_tasks(self, *ts)
 
     def add_head(self, t: Task, append: bool = True) -> None:
-        """Adds a task at the head of the workflow so the workflow start with the given task.
-
-        This sets the given task as a dependency of the starting tasks of the workflow.
-
-        Parameters
-        ----------
-        t: Task
-            The task to add to the head of the workflow.
-        append: bool = True
-            Whether to append the given task to the workflow.
-        """
-        if append:
-            self.add_task(t)
-
-        for template_task in self.dag_template.tasks:
-            if template_task.name != t.name:
-                if hasattr(template_task, 'dependencies'):
-                    template_task.dependencies.append(t.name)
-                else:
-                    setattr(template_task, 'dependencies', [t.name])
+        add_head(self, t, append=append)
 
     def add_tail(self, t: Task, append: bool = True) -> None:
-        """Adds a task as the tail of the workflow so the workflow ends with the given task.
-
-        This sets the given task's dependencies to all the tasks that are not listed as dependencies in the workflow.
-
-        Parameters
-        ----------
-        t: Task
-            The task to add to the tail of the workflow.
-        append: bool = True
-            Whether to append the given task to the workflow.
-        """
-        if append:
-            self.add_task(t)
-
-        dependencies = set()
-        task_name_to_task = dict()
-        for template_task in self.dag_template.tasks:
-            if hasattr(template_task, 'dependencies'):
-                dependencies.update(template_task.dependencies)
-            if template_task.name != t.name:
-                task_name_to_task[template_task.name] = template_task
-
-        # the tasks that are not listed as dependencies are "end tasks" i.e nothing follows after
-        # e.g if A -> B -> C then B.deps = [A] and C.deps = [B] but nothing lists C so C is "free"
-        free_tasks = set(task_name_to_task.keys()).difference(dependencies)
-        t.argo_task.dependencies = list(free_tasks)
-
-    def submit(self, namespace: Optional[str] = None) -> IoArgoprojWorkflowV1alpha1Workflow:
-        """Submits the workflow.
-
-        Notes
-        -----
-        This method is deprecated in favor of `workflow.create(...)`.
-        """
-        warnings.warn("`submit` is deprecated in favor of `create`", DeprecationWarning, stacklevel=2)
-        if namespace is None:
-            namespace = self.namespace
-        return self.service.submit(self.workflow, namespace)
+        add_tail(self, t, append=append)
 
     def create(self, namespace: Optional[str] = None) -> IoArgoprojWorkflowV1alpha1Workflow:
         """Creates the workflow"""
         if namespace is None:
             namespace = self.namespace
-        return self.service.submit(self.workflow, namespace)
+        return self.service.create(self.workflow, namespace)
+
+    def delete(self, namespace: Optional[str] = None) -> Tuple[object, int, dict]:
+        """Deletes the workflow"""
+        if namespace is None:
+            namespace = self.name
+        return self.service.delete(self.name)