+++ CHANGELOG.md
@@ -28,6 +28,20 @@ The general format is:
 - C from D
 
 ```
+
+# 3.0.0 - DATE (04/10/2022)
+
+### Added
+
+- update method to CronWorkflowService
+- ability to track workflow status
+- shared implementation of `add_head`, `add_tail`, `add_task`, `add_tasks`
+- task workflow template reference
+
+### Removed
+
+- UUID suffixes
+
 # 2.9.1 - DATE (04/11/2022)
 
 ### Fixed